third_party/WebKit/LayoutTests/fast/css/fontfaceset-invalid-family-names.html

<html>
<head>
<script src="../../resources/js-test.js"></script>
<style>
@font-face {
    font-family: sans-serif;
    src: local(Arial);
}

@font-face {
    font-family: serif;
    src: local(Arial);
}

@font-face {
    font-family: fantasy;
    src: local(Arial);
}

@font-face {
    font-family: cursive;
    src: local(Arial);
}

@font-face {
    font-family: monospace;
    src: local(Arial);
}

@font-face {
    font-family: inherit;
    src: local(Arial);
}

@font-face {
    font-family: initial;
    src: local(Arial);
}

@font-face {
    font-family: unset;
    src: local(Arial);
}

@font-face {
    font-family: default;
    src: local(Arial);
}
</style>
<script>
description('Tests that font-face with invalid family names are not added to FontFaceSet.');

function runTests() {
    shouldBe('document.fonts.size', '0');
}

if (document.fonts)
    runTests();
else
    testFailed('document.fonts does not exist');

</script>
</head>
<body>
</body>
</html>
