third_party/blink/web_tests/fast/css/fontfaceset-invalid-family-names-expected.txt

Tests that font-face with invalid family names are not added to FontFaceSet.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.fonts.size is 0
PASS successfullyParsed is true

TEST COMPLETE
